Time: 0 ms (100%), Space: 42.4 MB (54.94%) - LeetHub

+++ 0050-powx-n/0050-powx-n.java
@@ -1,7 +1,7 @@
 class Solution {
     public double myPow(double x, int n) {
         double ans = 1;
-        int nn = Math.abs(n);
+        long nn = Math.abs((long)n);
         while(nn>=1){
             if(nn%2==1){
                 ans=ans*x;
@@ -14,7 +14,7 @@ class Solution {
         if(n>=0){
             return ans;
         }else{
-            return (1.0)/ans;
+            return 1/ans;
         }
     }
 }
